changed type of confidence to double

+++ src/main/java/org/opennars/gui/graph/ImplicationGraph.java
@@ -38,7 +38,7 @@ public class ImplicationGraph extends SentenceGraph {
     
     @Override
     public boolean allow(final Sentence s) {        
-        float conf = s.truth.getConfidence();
+        double conf = s.truth.getConfidence();
         return conf > minConfidence;
     }
